refactor: 지역, 장르 데이터 분리 및 장르 추가
