+++ src/pages/ConcertList/constants/regionData.ts
@@ -1,0 +1,11 @@
+const regionList = [
+  "전국",
+  "서울/경기/인천",
+  "충청/대전/세종",
+  "경상/부산/대구/울산",
+  "전라/광주",
+  "강원",
+  "제주",
+];
+
+export default regionList;
