--- src/utils/constants
@@ -1,1 +1,0 @@
-상수 정의
